Move Azure App Services Modal components to subfolders
